+++ projects/js-packages/publicize-components/src/social-store/constants.ts
@@ -1,2 +1,6 @@
+// See projects/packages/publicize/src/jetpack-social-settings/class-settings.php
 export const SIG_SETTINGS_KEY = 'jetpack_social_image_generator_settings';
 export const UTM_ENABLED_KEY = 'jetpack_social_utm_settings';
+export const SOCIAL_NOTES_ENABLED_KEY = 'jetpack-social-note';
+export const SOCIAL_NOTES_CONFIG_KEY = 'jetpack_social_notes_config';
+export const SHOW_PRICING_PAGE_KEY = 'jetpack-social_show_pricing_page';