+++ projects/packages/publicize/src/jetpack-social-settings/class-settings.php
@@ -16,6 +16,8 @@ use Automattic\Jetpack\Publicize\Social_Image_Generator\Templates;
  * This class is used to get and update Jetpack_Social_Settings.
  * Currently supported features:
  *      - Social Image Generator
+ *      - UTM Settings
+ *      - Social Notes
  */
 class Settings {
 	/**
@@ -36,6 +38,17 @@ class Settings {
 	const DEFAULT_UTM_SETTINGS = array(
 		'enabled' => false,
 	);
+
+	const NOTES_CONFIG = 'notes_config';
+
+	const DEFAULT_NOTES_CONFIG = array(
+		'append_link' => true,
+	);
+
+	// Legacy named options.
+	const JETPACK_SOCIAL_NOTE_CPT_ENABLED   = 'jetpack-social-note';
+	const JETPACK_SOCIAL_SHOW_PRICING_PAGE  = 'jetpack-social_show_pricing_page';
+	const NOTES_FLUSH_REWRITE_RULES_FLUSHED = 'jetpack_social_rewrite_rules_flushed';
 
 	/**
 	 * Feature flags. Each item has 3 keys because of the naming conventions:
@@ -154,6 +167,58 @@ class Settings {
 			)
 		);
 
+		register_setting(
+			'jetpack_social',
+			self::JETPACK_SOCIAL_SHOW_PRICING_PAGE,
+			array(
+				'type'         => 'boolean',
+				'default'      => true,
+				'show_in_rest' => array(
+					'schema' => array(
+						'type' => 'boolean',
+					),
+				),
+			)
+		);
+
+		register_setting(
+			'jetpack_social',
+			self::JETPACK_SOCIAL_NOTE_CPT_ENABLED,
+			array(
+				'type'         => 'boolean',
+				'default'      => false,
+				'show_in_rest' => array(
+					'schema' => array(
+						'type' => 'boolean',
+					),
+				),
+			)
+		);
+
+		register_setting(
+			'jetpack_social',
+			self::OPTION_PREFIX . self::NOTES_CONFIG,
+			array(
+				'type'         => 'object',
+				'default'      => self::DEFAULT_NOTES_CONFIG,
+				'show_in_rest' => array(
+					'schema' => array(
+						'type'       => 'object',
+						'context'    => array( 'view', 'edit' ),
+						'properties' => array(
+							'append_link' => array(
+								'type' => 'boolean',
+							),
+							'link_format' => array(
+								'type' => 'string',
+								'enum' => array( 'full_url', 'shortlink', 'permashortcitation' ),
+							),
+						),
+					),
+				),
+			)
+		);
+
 		add_filter( 'rest_pre_update_setting', array( $this, 'update_settings' ), 10, 3 );
 	}
 
@@ -173,6 +238,24 @@ class Settings {
 	 */
 	public function get_utm_settings() {
 		return get_option( self::OPTION_PREFIX . self::UTM_SETTINGS, self::DEFAULT_UTM_SETTINGS );
+	}
+
+	/**
+	 * Get the social notes config.
+	 *
+	 * @return array The social notes config.
+	 */
+	public function get_social_notes_config() {
+		return get_option( self::OPTION_PREFIX . self::NOTES_CONFIG, self::DEFAULT_NOTES_CONFIG );
+	}
+
+	/**
+	 * Get if the social notes feature is enabled.
+	 *
+	 * @return bool
+	 */
+	public function is_social_notes_enabled() {
+		return get_option( self::JETPACK_SOCIAL_NOTE_CPT_ENABLED, false );
 	}
 
 	/**
@@ -253,10 +336,12 @@ class Settings {
 	 */
 	public function update_settings( $updated, $name, $value ) {
 
+		// Social Image Generator.
 		if ( self::OPTION_PREFIX . self::IMAGE_GENERATOR_SETTINGS === $name ) {
 			return $this->update_social_image_generator_settings( $value );
 		}
 
+		// UTM Settings.
 		if ( self::OPTION_PREFIX . self::UTM_SETTINGS === $name ) {
 			$current_utm_settings = $this->get_utm_settings();
 
@@ -265,6 +350,22 @@ class Settings {
 			}
 
 			return update_option( self::OPTION_PREFIX . self::UTM_SETTINGS, array_replace_recursive( $current_utm_settings, $value ) );
+		}
+
+		// Social Notes.
+		if ( self::JETPACK_SOCIAL_NOTE_CPT_ENABLED === $name ) {
+			// Delete this option, so the rules get flushed in maybe_flush_rewrite_rules when the CPT is registered.
+			delete_option( self::NOTES_FLUSH_REWRITE_RULES_FLUSHED );
+			return update_option( self::JETPACK_SOCIAL_NOTE_CPT_ENABLED, (bool) $value );
+		}
+		if ( self::OPTION_PREFIX . self::NOTES_CONFIG === $name ) {
+			$old_config = $this->get_social_notes_config();
+			$new_config = array_merge( $old_config, $value );
+			return update_option( self::OPTION_PREFIX . self::NOTES_CONFIG, $new_config );
+		}
+
+		if ( self::JETPACK_SOCIAL_SHOW_PRICING_PAGE === $name ) {
+			return update_option( self::JETPACK_SOCIAL_SHOW_PRICING_PAGE, (int) $value );
 		}
 
 		return $updated;