+++ projects/js-packages/boost-score-api/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.52] - 2025-01-23
+### Changed
+- Internal updates.
+
 ## [0.1.51] - 2025-01-20
 ### Changed
 - Updated package dependencies. [#41099]
@@ -219,6 +223,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Create package for the boost score bar API [#30781]
 
+[0.1.52]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.51...v0.1.52
 [0.1.51]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.50...v0.1.51
 [0.1.50]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.49...v0.1.50
 [0.1.49]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.48...v0.1.49